refactor(notifications): consolidate email formatting into email-channel.ts
Relocate getEmailSubject(), getEmailHtml(), and related helpers/types from src/lib/notification-formatting.ts to src/lib/notifications/channels/email-channel.ts.

- Update imports in src/app/api/unsubscribe/route.ts and tests.
- Delete src/lib/notification-formatting.ts.
- All notification and unsubscribe tests passing.

Closes PP-gk6
