+++ src/app/api/unsubscribe/route.ts
@@ -2,7 +2,7 @@ import { type NextRequest, NextResponse } from "next/server";
 import { eq } from "drizzle-orm";
 import { db } from "~/server/db";
 import { notificationPreferences } from "~/server/db/schema";
-import { verifyUnsubscribeToken } from "~/lib/notification-formatting";
+import { verifyUnsubscribeToken } from "~/lib/notifications/channels/email-channel";
 
 export function GET(request: NextRequest): NextResponse {
   const uid = request.nextUrl.searchParams.get("uid");
